Fix type for rank field

+++ src/main/java/io/algostrategy/client/coinmarketcap/domain/cryptocurrency/Cryptocurrency.java
@@ -20,7 +20,7 @@ public class Cryptocurrency {
 
     private Integer id;
 
-    private Double rank;
+    private Integer rank;
 
     private String name;
     private String symbol;
